Move scriptMode helper
