+++ navigator/src/epub/EpubNavigator.ts
@@ -19,7 +19,7 @@ import { IInjectableRule, IInjectablesConfig } from "../injection/Injectable.ts"
 import { IContentProtectionConfig, IKeyboardPeripheralsConfig } from "../Navigator.ts";
 import { NavigatorProtector, NAVIGATOR_SUSPICIOUS_ACTIVITY_EVENT } from "../protection/NavigatorProtector.ts";
 import { KeyboardPeripherals, NAVIGATOR_KEYBOARD_PERIPHERAL_EVENT } from "../peripherals/KeyboardPeripherals.ts";
-import { getScriptMode } from "./helpers/scriptMode.ts";
+import { getScriptMode } from "../helpers/scriptMode.ts";
 
 export type ManagerEventKey = "zoom";
