+++ navigator/src/epub/index.ts
@@ -3,4 +3,3 @@ export * from "./frame/index.ts";
 export * from "./fxl/index.ts";
 export * from "./preferences/index.ts";
 export * from "./css/index.ts";
-export { getScriptMode, type ScriptMode } from "./helpers/scriptMode.ts";
