+++ navigator/src/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from "./lineLength.ts";
 export * from './sML.ts';
+export * from './scriptMode.ts';
