+++ navigator/src/helpers/scriptMode.ts
